requirements.txt
beautifulsoup4==4.6.0
httplib2==0.9.2
requests==2.18.4
urllib3==1.22
python-dateutil==2.8.1
flask==1.1.2